Replace translation map with array of translation objects.
(Added in the scope of issue #209 "Test for duplicates".)

Duplicate keys in the map overwrite one another, so it is not possible
to test if a newly added translation is a duplicate of an existing one.

Replacing the map with an array of objects allows us to test for
duplicates.

`geeksayWord` is updated to handle the new format. The rest of the code
can remain the same.