Add package-lock to Git, and update .gitignore to ignore IDE configs
The package-lock should be in Git as it allows contributors to ensure
their dependencies match the ones used by the upstream CI.

IDE configs (such as .idea for IntelliJ IDEs) should be ignored by git
so first-time contributors can get started easily. This is especially
useful to new developers or those new to contributing to Open Source.